% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_grouping.R
\name{.perform_grouping}
\alias{.perform_grouping}
\title{Return the group number for each localization.}
\usage{
.perform_grouping(input, critDist, use_prog_bar = TRUE)
}
\arguments{
\item{input}{A numeric matrix where each row is a localization and each
column is a spatial axis.}

\item{critDist}{The critical distance for which localizations nearer than
this distance are deemed part of the same group.}

\item{use_prog_bar}{TRUE/FALSE indicating whether a progress bar should be
used. This is only available when run_parallel is FALSE.}
}
\description{
Group PALM/iPALM localizations based on their physical separation distance
}
\details{
PALM/iPALM data results in a list of spatial coordinates for fluorophore
localizations. This function groups nearby localizations if they are within
the provided critical distance from each other.
}
\author{
Zach Colburn
}
