% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blobs.R
\name{find_blobs}
\alias{find_blobs}
\title{Assign all neighboring pixels the same group number.}
\usage{
find_blobs(arr)
}
\arguments{
\item{arr}{A vector, matrix, or n-dimensional array where object-positive
elements are denoted by the value TRUE and background elements are denoted
by the value FALSE.}
}
\description{
Perform connected-component labeling to group continuous, thresholded
objects in n-dimensional arrays.

This function takes a vector, matrix, or n-dimensional array where each
element is TRUE if it corresponds to an object-positive index or FALSE if it
corresponds to a background index. An object of the same dimension as the
input is returned. All connected object indices take the value of their
group number and all background indices take the value NA.
}
\examples{
# Generate a random matrix.
set.seed(10)
mat <- matrix(runif(70), nrow = 7)

# Arbitrarily say that everything below 0.8 is background.
logical_mat <- mat > 0.8

# Find blobs.
find_blobs(logical_mat)

}
\author{
Zach Colburn
}
