% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_min_dists}
\alias{find_min_dists}
\title{For all points in matrix 1, return the distance and index of the
nearest point in matrix 2.}
\usage{
find_min_dists(mOne, mTwo)
}
\arguments{
\item{mOne}{A numeric matrix where each row is a localization and each
column is a spatial axis.}

\item{mTwo}{A matrix with the same number of columns as mOne.}
}
\description{
Find the shortest distance between each point in one data set and the points
in a second set.
}
\details{
This function determines the distance between every point in data set 1 and
the points in data set 2. Unlike this function's naive counterpart,
find_min_dists, this function divides the PALM/iPALM localization data
into blocks, opeartes on the data in each block, and then performs linking
operations on neighboring blocks.
}
\examples{
# Function call
\dontrun{find_min_dists(mOne, mTwo)}
}
\author{
Zach Colburn
}
