% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{identify_thresholded_objects}
\alias{identify_thresholded_objects}
\title{Assign all neighboring pixels the same group number.}
\usage{
identify_thresholded_objects(img, pixRange)
}
\arguments{
\item{img}{A thresholded matrix (where non-object pixels are assigned a
value of 0).}

\item{pixRange}{An integer number of pixels. Execution is faster when this
value is small. However, the value must be larger than the diameter of the
largest continuous object in the image.}
}
\description{
Group continuous, thresholded objects in rasterized objects.
}
\details{
This function takes a matrix corresponding to a thresholded image and
returns a matrix of the same size, where all adjacent, thesholded pixels
are the same integer corresponding to that object's cluster number.
}
\examples{
# Function call
identify_thresholded_objects(img, pixRange)

}
\author{
Zach Colburn
}
