% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shifter}
\alias{shifter}
\title{Identify the local shift between two temporally separated images.}
\usage{
shifter(t0, t1, xSeq, ySeq, searchBoxRadius, matchBoxRadius, shiftSkipper)
}
\arguments{
\item{t0}{A numeric matrix of intensity values for the first image.}

\item{t1}{A numeric matrix of intensity values for the second image.}

\item{xSeq}{A numeric vector of the x coordinates defining template centers.}

\item{ySeq}{A numeric vector of the y coordinates defining template centers.}

\item{searchBoxRadius}{An integer defining the search box radius.}

\item{matchBoxRadius}{An integer defining the radius of templates.}

\item{shiftSkipper}{An integer defining the number of pixels away the next
template to image comparison will be. This value must be 1 or greater.
Larger values will dramatically reduce execution time, but can reduce data
quality.}
}
\description{
Identify the local shift between two temporally separated images.
}
\details{
This function determines the local shift between two temporally separated
images. It does so by minimizing the squared difference between a local
template centered on the supplied peg coordinates and adjacent matrices of
the same dimension.
}
\examples{
# Function call
\dontrun{shifter(t0, t1, xSeq, ySeq, searchBoxRadius, matchBoxRadius, shiftSkipper)}
}
\author{
Zach Colburn
}
