% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_linker.R
\name{euclidean_linker}
\alias{euclidean_linker}
\title{Return the group number for each localization.}
\usage{
euclidean_linker(input, critDist)
}
\arguments{
\item{input}{A numeric matrix where each row is a localization and each
column is a spatial axis.}

\item{critDist}{The critical distance for which localizations nearer than
this distance are deemed part of the same group.}
}
\description{
Group PALM/iPALM localizations based on their physical separation distance
}
\details{
PALM/iPALM data results in a list of spatial coordinates for fluorophore
localizations. This function groups nearby localizations if they are within
the provided critical distance from each other.
}
\examples{
# Generate random data.
set.seed(10)
input <- as.matrix(data.frame(x=rnorm(10),y=rnorm(10)))

# Perform linking.
euclidean_linker(input, 0.4)

}
\author{
Zach Colburn
}
